pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-cli"
version = "0.1.0"
description = "Professional-grade CLI for OpenAI-compatible LLM routers"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.1.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"llm_cli" = ["py.typed"]
